Fix begin/end indices
Add `expr-idx` rule for expressions that are valid inside brackets,
disallowing statements, so `begin/end` can be used as indices.
This has the added benefit of removing the ambiguity between
comprehensions and for loops.

Some rules have to be parametrized, so they can use `expr` or `expr-idx`
depending of the context. This allows `begin/end` to be used inside
binary expressions, but not inside other expressions (like parentheses).

Overall, a very hacky solution... but it works.